Rework CSS HTML and JS, attach slider to grid
Scrapped much of my work because I felt using a the css grid function
would be more adaptable when the size of the etch-a-sketch needed to be
changed. The slider now changes the size of the grid and displays the
size. Now I just need to remember how I got the squares in the grid to
change when hovered over.